+++ misc/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+-   add `WEEK_START=sunday` environment variable
 -   fix live page fullscreen button
 
 ## `v0.3.5`